generate gfm with github too
