+++ test/fixtures/snapshots/gfm/github.html
@@ -1,0 +1,31 @@
+<!-- hash:Phspw7tq7d --><h1><a id="user-content-github-flavored-markdown" class="anchor" aria-hidden="true" tabindex="-1" href="#github-flavored-markdown"><span aria-hidden="true" class="octicon octicon-link"></span></a>GitHub Flavored Markdown</h1>
+<p>Read more: <a href="https://github.github.com/gfm/">https://github.github.com/gfm/</a></p>
+<h3><a id="user-content-autolink-literals" class="anchor" aria-hidden="true" tabindex="-1" href="#autolink-literals"><span aria-hidden="true" class="octicon octicon-link"></span></a>Autolink literals</h3>
+<p><a href="http://www.example.com" rel="nofollow">www.example.com</a>, <a href="https://example.com" rel="nofollow">https://example.com</a>, and <a href="mailto:contact@example.com">contact@example.com</a>.</p>
+<h3><a id="user-content-footnote" class="anchor" aria-hidden="true" tabindex="-1" href="#footnote"><span aria-hidden="true" class="octicon octicon-link"></span></a>Footnote</h3>
+<p>A note<sup><a href="#fn-1" id="user-content-fnref-1" data-footnote-ref="">1</a></sup></p>
+<h3><a id="user-content-strikethrough" class="anchor" aria-hidden="true" tabindex="-1" href="#strikethrough"><span aria-hidden="true" class="octicon octicon-link"></span></a>Strikethrough</h3>
+<p><del>one</del> or <del>two</del> tildes.</p>
+<h3><a id="user-content-table" class="anchor" aria-hidden="true" tabindex="-1" href="#table"><span aria-hidden="true" class="octicon octicon-link"></span></a>Table</h3>
+<table>
+<thead>
+<tr>
+<th>a</th>
+<th align="left">b</th>
+<th align="right">c</th>
+<th align="center">d</th>
+</tr>
+</thead>
+</table>
+<h2><a id="user-content-tasklist" class="anchor" aria-hidden="true" tabindex="-1" href="#tasklist"><span aria-hidden="true" class="octicon octicon-link"></span></a>Tasklist</h2>
+<ul>
+<li>[ ] to do</li>
+<li>[x] done</li>
+</ul>
+<section data-footnotes="">
+<ol>
+<li id="user-content-fn-1">
+<p>Big note. <a href="#fnref-1" data-footnote-backref="" aria-label="Back to reference 1">↩</a></p>
+</li>
+</ol>
+</section>
